Updated namespaces, and refactor Session from Entitites to new Entities folder
Removed Test.cs empty class from Entitites folder.
Updated target framework to net9.0 in DAL.csproj, enabled implicit usings and nullable reference types, and added necessary package references. Corrected namespace from DAL.Entitites to DAL.Entities across multiple files. Refactored Session.cs with updated class definition. Removed unused imports in _ViewImports.cshtml and Program.cs.
