+++ CoreGymBookingSystem/MainApp/Pages/_ViewImports.cshtml
@@ -1,5 +1,4 @@
 ﻿@using Microsoft.AspNetCore.Identity
 @using MainApp
-@using MainApp.Data
 @namespace MainApp.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
