+++ CoreGymBookingSystem/MainApp/Program.cs
@@ -1,7 +1,6 @@
 using DAL.DbContext;
 using DAL.Repositories;
 using DAL.Repositories.Interfaces;
-using MainApp.Data;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Service.Services;
